fix(FilePicker): Fix exported typings and cleanup file picker files
Fix typings being broken on productive build as importing types from `.vue` files does not work.
