+++ lib/components/types.ts
@@ -1,0 +1,35 @@
+/**
+ * @copyright Copyright (c) 2023 Ferdinand Thiessen <opensource@fthiessen.de>
+ *
+ * @author Ferdinand Thiessen <opensource@fthiessen.de>
+ *
+ * @license AGPL-3.0-or-later
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+import type { Node } from '@nextcloud/files'
+import type { AsyncComponent, Component } from 'vue'
+
+export interface IDialogButton {
+	label: string,
+	icon?: Component | AsyncComponent,
+	callback: () => void,
+	type?: 'primary' | 'secondary' | 'error' | 'warning' | 'success'
+}
+
+export interface IFilePickerButton extends Omit<IDialogButton, 'callback'> {
+	callback: (nodes: Node[]) => void
+}
