+++ lib/index.ts
@@ -1,8 +1,7 @@
-import { defineAsyncComponent } from 'vue'
-
 export { FilePicker, FilePickerType, FilePickerBuilder, getFilePickerBuilder } from './filepicker.js'
 export { TOAST_UNDO_TIMEOUT, TOAST_DEFAULT_TIMEOUT, TOAST_PERMANENT_TIMEOUT } from './toast.js'
 export { TOAST_ARIA_LIVE_OFF, TOAST_ARIA_LIVE_POLITE, TOAST_ARIA_LIVE_ASSERTIVE } from './toast.js'
 export { showMessage, showSuccess, showWarning, showInfo, showError, showUndo } from './toast.js'
 
-export const FilePickerVue = defineAsyncComponent(() => import('./components/FilePicker/FilePicker.vue'))
+export { spawnDialog } from './utils/dialogs.js'
+export { FilePickerVue } from './components/FilePicker/index.js'
