+++ lib/utils/l10n.ts
@@ -1,0 +1,34 @@
+/**
+ * @copyright Copyright (c) 2023 Ferdinand Thiessen <opensource@fthiessen.de>
+ *
+ * @author Ferdinand Thiessen <opensource@fthiessen.de>
+ * @author John Molakvoæ <skjnldsv@protonmail.com>
+ *
+ * @license AGPL-3.0-or-later
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+import { getGettextBuilder } from '@nextcloud/l10n/gettext'
+
+const gtBuilder = getGettextBuilder()
+	.detectLocale()
+
+__TRANSLATIONS__.map(data => gtBuilder.addTranslation(data.locale, data.json))
+
+const gt = gtBuilder.build()
+
+export const n = gt.ngettext.bind(gt) as typeof gt.ngettext
+export const t = gt.gettext.bind(gt) as typeof gt.gettext
